Adds the Aldous-Broder algorithm
It starts everything off by choosing one of the cells at random.
The random walk will begin at that cell.

To make sure the algorithm knows when everything has been visited,
it stores the number of unvisited cells in the grid.
We treat the starting cell as having been visited already.
Each time a new cell is visited, that value will is decremented,
and the loop continues until that value is zero”