Adds the Recursive Backtracker algorithm
This algorithm has to maintain a stack, which can potentially (worst-case)
contain every cell in the grid.
This suggests that the Recursive Backtracker can require up to twice as
much memory as Hunt-and-Kill algorithm.

The Recursive Backtracker, is guaranteed to visit each cell exactly twice
(once while carving, and once while backtracking),
this algorithm is generally going to be faster than Hunt-and-Kill.