Add movement by user input
I'm sure there will be a moment where this program will be forced to use 
Ncurses or Curses library, although that for now is far from the scope 
of my intentions. 

The main reason is that I don't want to introduce another library, 
secondly the library is poorly documented to be used in Ruby, one need 
to see original curses API and understand it. Thirdly, it doesn't play 
well with the Terminal text ouput. I'd have to create another way to 
print the game with it.