Update PNG output
This commit updated the PNG output.

It replaces the square shapes with circle shapes to represent
the starting point and the destination point.

It also increases the size of each cell from 20px to 128px,
this is purely to try and improve the printing quality.

I've also done a bit of a refactoring, trying to avoid too many
`if` statement conditions when possible.