Add spec file to MapUtils::Cell
I continue to believe that the distances method does not belong in this Cell object.

In fact, I start to believe that the following shouldn't be here:

- distances
- stairs?
- player?

This was probably a hack at the time to make it work, but in terms of
philosophical responsabilities the Cell object shouldn't need to know about
this conditions.